requirements.txt
aiohttp==3.6.3
appdirs==1.4.3
argcomplete==1.12.1
argon2-cffi==20.1.0
astroid==2.4.1
async-timeout==3.0.1
attrs==19.3.0
Automat==0.8.0
backcall==0.2.0
black==20.8b1
bleach==3.1.5
blessed==1.17.5
blinker==1.4
Brotli==1.0.9
certifi==2019.11.28
cffi==1.14.1
chardet==3.0.4
cli-helpers==2.0.1
click==7.1.2
cloud-init==20.3
colorama==0.4.3
command-not-found==0.3
commonmark==0.9.1
configobj==5.0.6
constantly==15.1.0
cryptography==2.9.2
dbus-python==1.2.16
decorator==4.4.2
defusedxml==0.6.0
delegator.py==0.1.1
discord.py==1.5.0
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
entrypoints==0.3
filelock==3.0.12
flake8==3.8.3
Flask==1.1.2
frida==12.11.0
frida-tools==8.1.0
fsb5==1.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.5.5
httplib2==0.14.0
httpx==0.12.1
hyperframe==5.2.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==16.10.1
iniconfig==1.1.1
ipykernel==5.3.4
ipython==7.16.1
ipython-genutils==0.2.0
isort==4.3.21
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.10.1
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
jupyter-client==6.1.6
jupyter-core==4.6.3
kaitaistruct==0.8
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
ldap3==2.7
litecli==1.4.1
lz4==3.1.0
MarkupSafe==1.1.0
mccabe==0.6.1
mistune==0.8.4
mitmproxy==5.2
more-itertools==4.2.0
multidict==4.7.6
mypy-extensions==0.4.3
nbconvert==5.6.1
nbformat==5.0.7
netifaces==0.10.4
notebook==6.1.1
oauthlib==3.1.0
objection==1.9.5
packaging==20.4
pandocfilters==1.4.2
parso==0.7.1
passlib==1.7.4
pathspec==0.8.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
pipx==0.15.6.0
pluggy==0.13.1
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.11.3
ptyprocess==0.6.0
publicsuffix2==2.20191221
py==1.9.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycodestyle==2.6.0
pycparser==2.20
pycryptodome==3.9.9
pyflakes==2.2.0
Pygments==2.6.1
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pylint==2.5.2
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyperclip==1.8.1
pyrsistent==0.15.5
pyserial==3.4
pytest==6.1.2
python-apt==2.0.0+ubuntu0.20.4.1
python-dateutil==2.8.1
python-debian===0.1.36ubuntu1
PyYAML==5.3.1
pyzmq==19.0.1
regex==2020.5.14
requests==2.22.0
requests-unixsocket==0.2.0
rfc3986==1.4.0
rich==9.1.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
SecretStorage==2.3.1
Send2Trash==1.5.0
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
sniffio==1.1.0
sortedcontainers==2.1.0
sos==4.0
SQLAlchemy==1.3.19
sqlparse==0.3.1
ssh-import-id==5.10
systemd-python==234
tabulate==0.8.7
terminado==0.8.3
terminaltables==3.1.0
testpath==0.4.4
texture2ddecoder==1.0.2
toml==0.10.1
tornado==6.0.4
traitlets==4.3.3
Twisted==18.9.0
typed-ast==1.4.1
typing-extensions==3.7.4.2
ubuntu-advantage-tools==20.3
ufw==0.36
unattended-upgrades==0.1
UnityPy==1.4.6.3
urllib3==1.25.8
urwid==2.0.1
userpath==1.4.1
virtualenv==20.0.20
wadllib==1.3.3
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
wsproto==0.15.0
yarl==1.5.1
zipp==1.0.0
zope.interface==4.7.1
zstandard==0.13.0
